Change the styles of the password fields
Previously, the validity or invalidity of the values of
the password fields was instantly notified to the user
only by changing the color of their borders. But now,
this border color change happens when the user clicks
on the 'Create Account' button and in addition, the invalidity of
their values is also displayed to the user as a message.